refactor(service): scope api properties + rename dynamic api properties
